Create GenerateAction
