+++ rules/rules_release/runner.js
@@ -1,16 +1,12 @@
-console.log(process.env.PWD);
 const { Command } = require("commander");
 const program = new Command();
-const getImpactedTargets = require("./lib/get-impacted-targets");
-const getConfig = require("./lib/get-config.js");
-const write = require("@changesets/write").default;
-const { rename, stat } = require("fs").promises;
-const fileExists = async (path) => !!(await stat(path).catch((e) => false));
 
 function collect(val, memo) {
   memo.push(val);
   return memo;
 }
+
+const GenerateAction = require("./lib/actions/GenerateAction");
 
 program
   .name("release-manager")
@@ -29,59 +25,33 @@ program
     "Additional args generate hashes command for bazel-diff"
   )
   .action(async (options) => {
-    const { config: configPaths } = program.opts();
-    const { bazelDiffPath, bazelDiffArgs } = options;
-    const config = await getConfig(configPaths);
-
-    console.log(config._releaseData);
-
-    const releaseLabels = config.releaseLabels();
-
-    const impactedTargets = getImpactedTargets({
-      bazelDiffPath,
-      bazelDiffArgs,
-      workspaceDir: config.workspaceDir(),
+    const action = new GenerateAction({
+      configPaths: program.opts().config,
+      bazelDiffPath: options.bazelDiffPath,
+      bazelDiffArgs: options.bazelDiffArgs,
     });
-
-    const changedReleaseLabels = impactedTargets.filter((target) => {
-      return releaseLabels.includes(target);
-    });
-
-    for (const changedReleaseLabel of changedReleaseLabels) {
-      const release = config.getReleaseByLabel(changedReleaseLabel);
-      const newFilePath = `${config.workspaceDir()}/.changeset/${
-        release.name
-      }.md`;
-
-      if (await fileExists(newFilePath)) {
-        console.log(`Skipping ${newFilePath} as it already exists`);
-        continue;
-      }
-
-      const changeset = {
-        summary: "A change",
-        releases: [{ name: release.name, type: "minor" }],
-      };
-      const uniqueId = await write(changeset, config.workspaceDir());
-      const oldFilePath = `${config.workspaceDir()}/.changeset/${uniqueId}.md`;
-
-      await rename(oldFilePath, newFilePath);
-      console.log(`Created changeset ${newFilePath}`);
-    }
+    await action.execute();
   });
 
 program
   .command("version")
   .description("Update version files based on changesets")
   .action(async () => {
-    const {
-      config: configPaths,
-      bazelDiffPath,
-      bazelDiffArgs,
-    } = program.opts();
-    const config = await getConfig(configPaths);
-
-    console.log(config._releaseData);
+    // const {
+    //   config: configPaths,
+    //   bazelDiffPath,
+    //   bazelDiffArgs,
+    // } = program.opts();
+    // const config = await getConfig(configPaths);
+    // console.log(config._releaseData);
+    console.log("version");
   });
 
 program.parse();
+
+// cli program
+// generate action
+// release repository: get releases / get changed releases
+// target repository: get impacted targets
+// changeset repository: write changesets
+// version action
